[#11] Create lp_problem module
