+++ conference_scheduler/tests/conftest.py
@@ -2,10 +2,10 @@ import pytest
 import numpy as np
 from conference_scheduler.resources import (
     Person, Room, Slot, Session, EventType, Event, Role, Demand,
-    Unavailability, ScheduledItem
+    ScheduledItem, Shape
 )
 from conference_scheduler import scheduler
-from conference_scheduler import parameters
+from conference_scheduler.lp_problem import utils as lpu
 
 
 @pytest.fixture(scope="module")
@@ -108,32 +108,32 @@ def demand(events):
 
 @pytest.fixture(scope='module')
 def shape(events, slots):
-    return parameters.Shape(len(events), len(slots))
+    return Shape(len(events), len(slots))
 
 
 @pytest.fixture(scope='module')
 def tag_array(events):
-    return parameters.tag_array(events)
+    return lpu.tag_array(events)
 
 
 @pytest.fixture(scope='module')
 def session_array(sessions):
-    return parameters.session_array(sessions)
+    return lpu.session_array(sessions)
 
 
 @pytest.fixture(scope='module')
 def slot_availability_array(events, slots):
-    return parameters.slot_availability_array(events, slots)
+    return lpu.slot_availability_array(events, slots)
 
 
 @pytest.fixture(scope='module')
 def event_availability_array(events, slots):
-    return parameters.event_availability_array(events)
+    return lpu.event_availability_array(events)
 
 
 @pytest.fixture(scope='module')
 def X(shape):
-    return parameters.variables(shape)
+    return lpu.variables(shape)
 
 
 @pytest.fixture(scope='module')
