+++ conference_scheduler/tests/lp_problem/test_constraints.py
@@ -1,72 +1,10 @@
-from conference_scheduler import parameters
 import numpy as np
-
-
-def test_tag_array(events):
-    tag_array = parameters.tag_array(events)
-    assert np.array_equal(tag_array, np.array([[1, 0], [1, 1], [0, 1]]))
-
-
-def test_session_array(sessions):
-    session_array = parameters.session_array(sessions)
-    assert np.array_equal(session_array, np.array([
-        [1, 1, 1, 0, 0, 0, 0],
-        [0, 0, 0, 1, 1, 0, 0],
-        [0, 0, 0, 0, 0, 1, 0],
-        [0, 0, 0, 0, 0, 0, 1]
-    ]))
-
-
-def test_slot_availability_array(events, slots):
-    slot_availability_array = parameters.slot_availability_array(events, slots)
-    assert np.array_equal(slot_availability_array, np.array([
-        [0, 0, 1, 1, 1, 1, 1],
-        [1, 1, 0, 0, 1, 1, 1],
-        [1, 1, 1, 1, 1, 1, 1]
-    ]))
-
-
-def test_event_availability_array(events):
-    event_availability_array = parameters.event_availability_array(events)
-    assert np.array_equal(event_availability_array, np.array([
-        [1, 0, 1, ],
-        [0, 1, 1, ],
-        [1, 1, 1, ]
-    ]))
-
-
-def test_slots_overlap(slots):
-    assert parameters.slots_overlap(slots[0], slots[1]) is False
-    assert parameters.slots_overlap(slots[0], slots[2]) is False
-    assert parameters.slots_overlap(slots[0], slots[3]) is False
-    assert parameters.slots_overlap(slots[0], slots[4]) is False
-    assert parameters.slots_overlap(slots[0], slots[6]) is False
-    assert parameters.slots_overlap(slots[1], slots[6]) is False
-    assert parameters.slots_overlap(slots[1], slots[4]) is False
-    assert parameters.slots_overlap(slots[6], slots[1]) is False
-    assert parameters.slots_overlap(slots[1], slots[6]) is False
-
-    assert parameters.slots_overlap(slots[0], slots[0]) is True
-    assert parameters.slots_overlap(slots[5], slots[0]) is True
-    assert parameters.slots_overlap(slots[0], slots[5]) is True
-    assert parameters.slots_overlap(slots[1], slots[5]) is True
-    assert parameters.slots_overlap(slots[5], slots[1]) is True
-    assert parameters.slots_overlap(slots[6], slots[6]) is True
-
-
-def test_concurrent_slots(slots):
-    slots = list(parameters.concurrent_slots(slots))
-    assert slots == [(0, 5), (1, 5), (2, 6), (3, 6), (4, 6)]
-
-
-def test_variables(shape):
-    X = parameters.variables(shape)
-    assert len(X) == 21
+from conference_scheduler.lp_problem import constraints as lpc
 
 
 def test_schedule_all_events(shape, X):
     constraints = [
-        c.condition for c in parameters._schedule_all_events(shape, X)]
+        c.condition for c in lpc._schedule_all_events(shape, X)]
     assert len(constraints) == 3
 
 
@@ -78,7 +16,7 @@ def test_schedule_all_events_fails_np(shape):
         [0, 0, 0, 0, 0, 0, 0]
     ])
     constraints = [
-        c.condition for c in parameters._schedule_all_events(shape, X)]
+        c.condition for c in lpc._schedule_all_events(shape, X)]
     assert not all(constraints)
 
 
@@ -90,13 +28,13 @@ def test_schedule_all_events_pass_np(shape):
         [0, 1, 0, 0, 0, 0, 0]
     ])
     constraints = [
-        c.condition for c in parameters._schedule_all_events(shape, X)]
+        c.condition for c in lpc._schedule_all_events(shape, X)]
     assert all(constraints)
 
 
 def test_max_one_event_per_slot(shape, X):
     constraints = [
-        c.condition for c in parameters._max_one_event_per_slot(shape, X)]
+        c.condition for c in lpc._max_one_event_per_slot(shape, X)]
     assert len(constraints) == 7
 
 
@@ -108,7 +46,7 @@ def test_max_one_events_per_slot_fail_np(shape):
         [0, 1, 0, 0, 0, 0, 0]
     ])
     constraints = [
-        c.condition for c in parameters._max_one_event_per_slot(shape, X)]
+        c.condition for c in lpc._max_one_event_per_slot(shape, X)]
     assert not all(constraints)
 
 
@@ -120,28 +58,12 @@ def test_max_one_events_per_slot_pass_np(shape):
         [0, 1, 0, 0, 0, 0, 0]
     ])
     constraints = [
-        c.condition for c in parameters._max_one_event_per_slot(shape, X)]
+        c.condition for c in lpc._max_one_event_per_slot(shape, X)]
     assert all(constraints)
 
 
-def test_slots_in_session(session_array):
-    assert np.array_equal(parameters._slots_in_session(0, session_array),
-                          np.array([0, 1, 2]))
-    assert np.array_equal(parameters._slots_in_session(3, session_array),
-                          np.array([6]))
-
-
-def test_events_with_diff_tags(tag_array):
-    assert np.array_equal(parameters._events_with_diff_tag(0, tag_array),
-                          np.array([2]))
-    assert np.array_equal(parameters._events_with_diff_tag(1, tag_array),
-                          np.array([]))
-    assert np.array_equal(parameters._events_with_diff_tag(2, tag_array),
-                          np.array([0]))
-
-
 def test_events_in_session_share_a_tag(session_array, tag_array, X):
-    constraints = [c for c in parameters._events_in_session_share_a_tag(
+    constraints = [c for c in lpc._events_in_session_share_a_tag(
         session_array, tag_array, X)]
     assert len(constraints) == 16
 
@@ -154,7 +76,7 @@ def test_events_in_session_share_a_tag_fails_np(session_array, tag_array):
         [0, 1, 0, 0, 0, 0, 0]
     ])
     constraints = [
-        c.condition for c in parameters._events_in_session_share_a_tag(
+        c.condition for c in lpc._events_in_session_share_a_tag(
             session_array, tag_array, X)]
     assert not all(constraints)
 
@@ -166,14 +88,14 @@ def test_events_in_session_share_a_tag_passes_np(session_array, tag_array):
         [0, 0, 0, 0, 0, 0, 1],
         [0, 0, 0, 1, 0, 0, 0]
     ])
-    test = all(parameters._events_in_session_share_a_tag(session_array,
-                                                         tag_array, X))
+    test = all(lpc._events_in_session_share_a_tag(
+        session_array, tag_array, X))
     assert test is True
 
 
 def test_events_available_in_scheduled_slot(slot_availability_array, X):
     constraints = [
-        c for c in parameters._events_available_in_scheduled_slot(
+        c for c in lpc._events_available_in_scheduled_slot(
             slot_availability_array, X)]
     assert len(constraints) == 21
 
@@ -186,7 +108,7 @@ def test_events_available_in_scheduled_slot_fails_np(slot_availability_array):
         [0, 1, 0, 0, 0, 0, 0]
     ])
     constraints = [
-        c.condition for c in parameters._events_available_in_scheduled_slot(
+        c.condition for c in lpc._events_available_in_scheduled_slot(
             slot_availability_array, X)]
     assert all(constraints) is False
 
@@ -200,7 +122,7 @@ def test_events_available_in_scheduled_slot_passes_np(slot_availability_array):
     ])
     constraints = [
         c.condition
-        for c in parameters._events_available_in_scheduled_slot(
+        for c in lpc._events_available_in_scheduled_slot(
             slot_availability_array, X)]
     assert all(constraints) is True
 
@@ -209,7 +131,7 @@ def test_events_available_during_other_events(
     event_availability_array, slots, X
 ):
     constraints = [
-        c for c in parameters._events_available_during_other_events(
+        c for c in lpc._events_available_during_other_events(
             event_availability_array, slots, X)]
     assert len(constraints) == 45
 
@@ -224,7 +146,7 @@ def test_events_available_during_other_events_fails_np(
         [0, 1, 0, 0, 0, 0, 0]
     ])
     constraints = [
-        c.condition for c in parameters._events_available_during_other_events(
+        c.condition for c in lpc._events_available_during_other_events(
             event_availability_array, slots, X)]
     assert all(constraints) is False
 
@@ -239,7 +161,7 @@ def test_events_available_during_other_events_pass_np(
         [0, 1, 0, 0, 0, 0, 0]
     ])
     constraints = [
-        c.condition for c in parameters._events_available_during_other_events(
+        c.condition for c in lpc._events_available_during_other_events(
             event_availability_array, slots, X)]
     assert all(constraints) is True
 
@@ -249,7 +171,7 @@ def test_constraints(
     event_availability_array, X
 ):
     constraints = [
-        c for c in parameters.constraints(
+        c for c in lpc.constraints(
             events, slots, session_array, tag_array, slot_availability_array,
             event_availability_array, X)]
     assert len(constraints) == 92
